Add ITypeEqualityComparer to override equality checks.
Some types do not correctly check for deep equality. If one of those
types is used as a component the visual debugger's inspector will not be
able to understand whether a change was applied. An example of this is
Unity's `AnimationCurve`.

This change adds a new interface type to declare type equality comparers
which are used instead of calling `Equals` on the values when changing
something in the inspector in Unity. An implementation of a type
equality comparer for `AnimationCurve` is provided, too.